split out voices tab
