+++ src/history_tab/main.py
@@ -8,14 +8,12 @@ from src.history_tab.collections_directories_atom import (
     get_collections,
 )
 from src.history_tab.get_wav_files import (
-    get_npz_files_voices,
     get_wav_files,
     get_wav_files_img,
 )
 from src.history_tab.delete_generation_cb import delete_generation_cb
 from src.history_tab.save_to_favorites import save_to_collection, save_to_favorites
 from src.history_tab.open_folder import open_folder
-from src.bark.get_audio_from_npz import get_audio_from_npz
 
 
 def _get_row_index(evt: gr.SelectData):
@@ -281,95 +279,3 @@ def create_collection_ui(directories_state):
         inputs=[new_collection_name],
         outputs=[directories_state, create_collection_button],
     )
-
-
-def voices_tab(register_use_as_history_button, directory="voices"):
-    with gr.Tab(directory.capitalize()) as voices_tab, gr.Row():
-        with gr.Column():
-            with gr.Row():
-                button_output = gr.Button(value=f"Open {directory} folder")
-            button_output.click(lambda: open_folder(directory))
-
-            datatypes = ["date", "str", "str", "str", "str"]
-            headers = [
-                "Date and Time",
-                directory.capitalize(),
-                "When",
-                "Hash",
-                "Filename",
-            ]
-
-            voices_list = gr.Dataframe(
-                value=get_npz_files_voices(),
-                interactive=False,
-                datatype=datatypes,
-                col_count=len(datatypes),
-                max_cols=len(datatypes),
-                headers=headers,
-                #  elem_classes="file-list"
-            )
-        with gr.Column():
-            audio = gr.Audio(visible=True, type="numpy", label="Fine prompt audio")
-            voice_file_name = gr.Textbox(label="Voice file name", value="")
-            new_voice_file_name = gr.Textbox(label="New voice file name", value="")
-
-            with gr.Row():
-                delete_voice_button = gr.Button(value="Delete voice", variant="stop")
-                use_voice_button = gr.Button(value="Use voice", variant="primary")
-                rename_voice_button = gr.Button(value="Rename voice")
-
-    def delete_voice(voice_file_name):
-        os.remove(voice_file_name)
-        return {
-            delete_voice_button: gr.Button.update(value="Deleted"),
-            voices_list: update_voices_tab(),
-        }
-
-    def rename_voice(voice_file_name, new_voice_file_name):
-        shutil.move(voice_file_name, new_voice_file_name)
-        return {
-            rename_voice_button: gr.Button.update(value="Renamed"),
-            voices_list: update_voices_tab(),
-        }
-
-    rename_voice_button.click(
-        fn=rename_voice,
-        inputs=[voice_file_name, new_voice_file_name],
-        outputs=[rename_voice_button, voices_list],
-    )
-    register_use_as_history_button(
-        use_voice_button,
-        voice_file_name,
-    )
-    delete_voice_button.click(
-        fn=delete_voice,
-        inputs=[voice_file_name],
-        outputs=[delete_voice_button, voices_list],
-    )
-
-    def select(_list_data, evt: gr.SelectData):
-        filename_npz = _get_filename(_list_data, _get_row_index(evt))
-        return {
-            voice_file_name: gr.Textbox.update(value=filename_npz),
-            new_voice_file_name: gr.Textbox.update(value=filename_npz),
-            delete_voice_button: gr.Button.update(value="Delete"),
-            rename_voice_button: gr.Button.update(value="Rename"),
-            audio: gr.Audio.update(value=get_audio_from_npz(filename_npz)),
-        }
-
-    outputs = [
-        voice_file_name,
-        new_voice_file_name,
-        delete_voice_button,
-        rename_voice_button,
-        audio,
-    ]
-
-    voices_list.select(
-        fn=select, inputs=[voices_list], outputs=outputs, preprocess=False
-    )
-
-    def update_voices_tab():
-        return gr.List.update(value=get_npz_files_voices())
-
-    voices_tab.select(fn=update_voices_tab, outputs=[voices_list])
